add waitForIdentified method
